refactor: rename Utilities to Game and update namespaces for Emote and FloatText managers
